requirements.txt
# Bibliotecas principales
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
scipy==1.13.1

# Visualización de datos
#matplotlib==3.9.1  # Actualiza a una versión compatible
seaborn==0.13.2
plotly==5.23.0

# Procesamiento de lenguaje natural
nltk==3.8.1
spacy==3.8.4
transformers==4.39.3

# Machine Learning y Deep Learning
tensorflow==2.17.0
torch==2.3.1

# Streamlit y dependencias relacionadas
streamlit==1.37.0
fastapi==0.109.2
httpx==0.27.0

# Otras dependencias
tqdm==4.66.2
python-dotenv==1.0.1

### Faltan whisper and whisperx nltk spacy 
###